edited the dataloader to load researchpapers and articles also fixed issuesin endpoints and add two more endpoints in interactions added authorization to some endpoints too

+++ src/main/java/edu/bethlehem/scinexus/Article/ArticleController.java
@@ -5,27 +5,31 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.http.*;
-
+import org.springframework.security.core.Authentication;
 import org.springframework.hateoas.*;
 import org.springframework.web.bind.annotation.*;
 
+import edu.bethlehem.scinexus.Config.JwtService;
+import edu.bethlehem.scinexus.Config.UserDetailsImpl;
+import edu.bethlehem.scinexus.User.User;
+import edu.bethlehem.scinexus.User.UserNotFoundException;
+import edu.bethlehem.scinexus.User.UserRepository;
+import lombok.RequiredArgsConstructor;
+
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/articles")
 public class ArticleController {
 
   private final ArticleRepository repository;
+  private final UserRepository userRepository;
   private final ArticleModelAssembler assembler;
-
-  ArticleController(ArticleRepository repository, ArticleModelAssembler assembler) {
-    this.repository = repository;
-    this.assembler = assembler;
-  }
 
   @GetMapping("/{articleId}")
   EntityModel<Article> one(@PathVariable Long articleId) {
 
     Article article = repository.findById(articleId)
-        .orElseThrow(() -> new ArticleNotFoundException(articleId,HttpStatus.NOT_FOUND));
+        .orElseThrow(() -> new ArticleNotFoundException(articleId, HttpStatus.NOT_FOUND));
 
     return assembler.toModel(article);
   }
@@ -39,7 +43,13 @@ public class ArticleController {
   }
 
   @PostMapping()
-  ResponseEntity<?> newArticle(@RequestBody Article newArticle) {
+  ResponseEntity<?> newArticle(@RequestBody Article newArticle,
+      Authentication authentication) {
+
+    User userObj = (User) authentication.getPrincipal();
+    User user = userRepository.findById(userObj.getId())
+        .orElseThrow(() -> new UserNotFoundException(userObj.getId().toString(), HttpStatus.NOT_FOUND));
+    newArticle.setPublisher(user);
 
     EntityModel<Article> entityModel = assembler.toModel(repository.save(newArticle));
 
@@ -47,34 +57,41 @@ public class ArticleController {
   }
 
   @PutMapping("/{id}")
-  ResponseEntity<?> editArticle(@RequestBody Article newArticle, @PathVariable Long id) {
+  ResponseEntity<?> editArticle(@RequestBody Article newArticle, @PathVariable Long id,
+      Authentication authentication) {
 
-    return repository.findById(id)
-        .map(article -> {
-          article.setContent(newArticle.getContent());
-          article.setDescription(newArticle.getDescription());
-          article.setSubject(newArticle.getSubject());
-          article.setTitle(newArticle.getTitle());
+    User user = (User) authentication.getPrincipal();
 
-          article.setPublisher(newArticle.getPublisher());
-          article.setVisibility(newArticle.getVisibility());
-          article.setContributors(newArticle.getContributors());
-          EntityModel<Article> entityModel = assembler.toModel(repository.save(article));
-          return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
-        })
-        .orElseGet(() -> {
-          newArticle.setId(id);
-          EntityModel<Article> entityModel = assembler.toModel(repository.save(newArticle));
-          return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
-        });
+    Article article = repository.findById(id).orElseThrow(() -> new ArticleNotFoundException(id, HttpStatus.NOT_FOUND));
+    if (!article.getContributors().contains(user) && !article.getPublisher().equals(user))
+      // return forbidden
+      throw new ArticleNotFoundException(id, HttpStatus.NOT_FOUND);
+
+    article.setContent(newArticle.getContent());
+    article.setDescription(newArticle.getDescription());
+    article.setSubject(newArticle.getSubject());
+    article.setTitle(newArticle.getTitle());
+    article.setPublisher(newArticle.getPublisher());
+    article.setVisibility(newArticle.getVisibility());
+    article.setContributors(newArticle.getContributors());
+
+    EntityModel<Article> entityModel = assembler.toModel(repository.save(article));
+    return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
+
   }
 
   @PatchMapping("/{id}")
-  public ResponseEntity<?> updateUserPartially(@PathVariable(value = "id") Long articleId,
-      @RequestBody Article newArticle) {
-    Article article = repository.findById(articleId)
-        .orElseThrow(() -> new ArticleNotFoundException(articleId,HttpStatus.NOT_FOUND));
+  public ResponseEntity<?> updateUserPartially(@PathVariable(value = "id") Long id,
+      @RequestBody Article newArticle,
+      Authentication authentication) {
 
+    // this section checks if the user have access to the article
+    User user = (User) authentication.getPrincipal();
+    Article article = repository.findById(id).orElseThrow(() -> new ArticleNotFoundException(id, HttpStatus.NOT_FOUND));
+    if (!article.getContributors().contains(user) && !article.getPublisher().equals(user))
+      // return forbidden
+      throw new ArticleNotFoundException(id, HttpStatus.NOT_FOUND);
+    // //////////////////////////////////////////////////////////////////////
     if (newArticle.getContent() != null)
       article.setContent(newArticle.getContent());
     if (newArticle.getDescription() != null)
@@ -96,9 +113,16 @@ public class ArticleController {
   }
 
   @DeleteMapping("/{id}")
-  ResponseEntity<?> deleteArticle(@PathVariable Long id) {
+  ResponseEntity<?> deleteArticle(@PathVariable Long id,
+      Authentication authentication) {
 
-    Article article = repository.findById(id).orElseThrow(() -> new ArticleNotFoundException(id,HttpStatus.NOT_FOUND));
+    // this section checks if the user have access to the article
+    User user = (User) authentication.getPrincipal();
+    Article article = repository.findById(id).orElseThrow(() -> new ArticleNotFoundException(id, HttpStatus.NOT_FOUND));
+    if (!article.getContributors().contains(user) && !article.getPublisher().equals(user))
+      // return forbidden
+      throw new ArticleNotFoundException(id, HttpStatus.NOT_FOUND);
+    // //////////////////////////////////////////////////////////////////////
 
     repository.delete(article);
